fix(custom_watch): carry the u64 pages mask into the new preview call site
Integration seam between two parallel worktrees: one widened
statusbar::page_indicator's mask to u64 (page 33 overflowed the u32), the
other added a RouteElev preview test passing u32::MAX. Both merged
cleanly and the result did not compile.

+++ apps/custom_watch/render/src/preview.rs
@@ -445,7 +445,7 @@ fn preview_route_elev_page() {
     draw_face(&mut fb, Page::RouteElev, Some(&snap), None);
     widgets::draw_page_indicator(
         &mut fb,
-        watch_core::statusbar::page_indicator(Page::RouteElev, u32::MAX),
+        watch_core::statusbar::page_indicator(Page::RouteElev, u64::MAX),
     );
     widgets::draw_route_elev_overlay(&mut fb, &snap);
     show("route elevation: course profile + position marker", &fb);
